Attempt 3 - Map not loading
Based on the error message I got on the last attempt, I switched out the API to the one I was able to generate with my Google account.

However, I cannot manage this API because this feature is stuck behind a paywall that asks for information I don't want to fill out.

I will try to generate a different API and see what happens.

I also removed the link to CSS found in the head, because one of the errors was that the stylesheet was not found. Since I am trying to get this to work without copying any CSS, I will continue to work from this point forward.