Attempt 3 - Stuck behind paywall
I made some changes according to Google's documentation on how to use the API & load up the map.

I didn't get anything, but the console states that the error is "BillingNotEnabledMapError"

Again, I would like to avoid filling out the paywall, because it asks for information such as organization address, billing address, and credit card info, and I do not want to risk providing different types of information unless I truly have to.