More Progress on Image Slider
I've been struggling to figure out how to get the slider to work.

Currently, I have a situation where the buttons to select a different image either add or remove images and stack them underneath the main image.

I am most positive that this is a CSS issue rather than a JS issue because everything seems to be behaving correctly (according to the console) but the image won't replace the one before it.