Update 7 - Only BillingError
I've edited the HTML for the maps page and the only error im getting is the billing error, which I am not sure what to do for.

I will have to see what to do with the js file as there is a possibility i might need to change it.