All Project Deliverables Finalized
Finished the media queries, and this should finalize my submission.

The Page I chose to modify was the PC Builds page. The elements that change are the wrapper background, text alignment and colors of the intro paragraph, 4 images, and the navigation buttons.

I edited the CSS for my image slider and made the buttons bigger, and the height of the image container larger to accommodate all images.

For the map, I took the professor's exact demo files, and modified them to my location and liking. I wanted to experiment with adding shapes to the map, but I had multiple attempts that didn't work, and I decided to move on.

All in all, this submission should address all the requirements for project 3.